requirements.txt
numpy>=1.24
PyOpenGL>=3.1.7
pygame>=2.5.0
pytest>=7.4.0
